refactor: CardListWithPaginator
